+++ src/shared/ui/form/FormContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import { FormContextValue } from './types';
+
+export const FormContext = createContext<FormContextValue<Record<string, any>> | null>(null);
