+++ src/shared/ui/form/index.ts
@@ -1,0 +1,4 @@
+export { Form } from './Form';
+export { FormItem } from './FormItem';
+export { FormSubmit, FormReset } from './FormActions';
+export type { FormProps, FormItemProps, FormSubmitProps, FormResetProps } from './types';
